+++ django_geoip/views.py
@@ -2,7 +2,7 @@
 from django import http
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist
-from django_geoip.base import LocationStorage
+from django_geoip.base import storage_class
 from django_geoip.utils import get_class
 
 def set_location(request):
@@ -27,7 +27,7 @@ def set_location(request):
         if location_id:
             try:
                 location = get_class(settings.GEOIP_LOCATION_MODEL).objects.get(pk=location_id)
-                LocationStorage(request=request, response=response).set(location=location, force=True)
+                storage_class(request=request, response=response).set(location=location, force=True)
             except (ValueError, ObjectDoesNotExist):
                 pass
     return response