+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 * Fixed middleware behavior when request.location is None
+* Added ``GEOIP_STORAGE_CLASS`` setting to override default user location storage
 
 
 0.2 (2012-01-20)